Per @thiamsantos's requests: * moves unit tests from ./src/tests to ./test * moved test helpers from ./test/test.js to ./test/test-helpers.js
